move ACTION_METHODS to constants file to remove circular dependency
