+++ src/constants/actionlist.consts.ts
@@ -1,0 +1,32 @@
+import { MERCADO_PAGO_ICON } from '@/assets'
+import { METAMASK_LOGO, TRUST_WALLET_SMALL_LOGO } from '@/assets/wallets'
+import binanceIcon from '@/assets/exchanges/binance.svg'
+import { Method } from '@/components/Common/ActionList'
+
+export const ACTION_METHODS: Method[] = [
+    {
+        id: 'bank',
+        title: 'Bank',
+        description: 'EUR, USD, ARS (more coming soon)',
+        icons: [
+            'https://flagcdn.com/w160/ar.png',
+            'https://flagcdn.com/w160/de.png',
+            'https://flagcdn.com/w160/us.png',
+        ],
+        soon: false,
+    },
+    {
+        id: 'mercadopago',
+        title: 'Mercado Pago',
+        description: 'Instant transfers',
+        icons: [MERCADO_PAGO_ICON],
+        soon: true,
+    },
+    {
+        id: 'exchange-or-wallet',
+        title: 'Exchange or Wallet',
+        description: 'Binance, Coinbase, Metamask and more',
+        icons: [binanceIcon, TRUST_WALLET_SMALL_LOGO, METAMASK_LOGO],
+        soon: false,
+    },
+]
